Feat : save products in localStorage
Avoid calling everytime we reload the page
Call api every 10 minutes
