+++ src/___tests___/Home.spec.js
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "vitest";
 
 import { mount } from "@vue/test-utils";
-import Home from "@/components/Home.vue";
+import Home from "@/pages/Home.vue";
 
 describe("Home", ()=>{
     it("renders proprely", ()=>{
