--- src/components/Home.vue
@@ -1,46 +1,0 @@
-<script setup async>
-import { ref, onMounted } from 'vue';
-import { useRoute } from 'vue-router'
-import { computed } from 'vue'
-import PrimaryButton from './PrimaryButton.vue';
-import ProductCard from './ProductCard.vue';
-import SuccessAlert from './SuccessAlert.vue';
-import DangerAlert from './DangerAlert.vue';
-
-const route = useRoute()
-let successMessage = ref('')
-if (route !== undefined){
-    successMessage.value =  route.query.success
-}
-const errorMessages = ref([])
-
-
-const products = ref([]);
-
-onMounted(async () => {
-    try {
-        const res = await fetch('https://fakestoreapi.com/products');
-        const data = await res.json();
-        products.value = data;
-    } catch (error) {
-        errorMessages.values.push("Erreur lors de la récupération des produits :", error);
-    }
-});
-</script>
-
-<template>
-    <div class="m-4">
-        <SuccessAlert v-if="successMessage" :text="successMessage" />
-        <div v-if="errorMessages">
-            <DangerAlert v-for="error of errorMessages" :text="error" />
-        </div>
-        
-        <header class="sm:flex justify-between">
-            <h1 class="text-xl text-(--color-heading) font-bold">Répertoire des produits</h1>
-            <PrimaryButton type="button" text="Ajouter" action="add"/>
-        </header>
-        <section class="md:grid md:grid-cols-3 md:gap-3">
-            <ProductCard v-for="product in products" :key="product.id" :img="product.image" :price="product.price" :label="product.title" :description="product.description" :category="product.category"/>
-        </section>
-    </div>
-</template>
